refactor(covid): more component reorganization
