+++ apps/covid/src/components/MyContext/MyContext.svelte
@@ -8,8 +8,7 @@
     Text,
   } from '@studiobear/designspek-components'
   import { GetContextFab, RevGeocode } from '../util'
-  import Modal from '../Modal.svelte'
-  import Loading from '../Loading.svelte'
+  import { Modal, Loading } from '../shared'
   import OverviewBoxLocal from './OverviewBoxLocal.svelte'
   import { insertCommas, getAddress, setStorageItem } from '../../libs'
   import { storeUserPrefs, lfUserPrefs } from '../../stores/userPrefs'
