+++ apps/covid/src/components/Stats/TableSubRegional.svelte
@@ -21,7 +21,7 @@
     internalizeCountryName,
     percentChange,
   } from '../../libs'
-  import Loading from '../Loading.svelte'
+  import {Loading} from '../shared'
   import {
     H2,
     H4,
