+++ apps/covid/src/components/index.js
@@ -1,17 +1,8 @@
 import Nav from './Nav.svelte'
-import SSR from './styleSSR.svelte'
-import Icons from './Icons.svelte'
-import Icon from './Icon.svelte'
-import Overlay from './Overlay.svelte'
-import Loading from './Loading.svelte'
+import OverviewBoxGlobal from './OverviewBoxGlobal.svelte'
 
 export * from './MyContext'
 export * from './Stats'
-export {
-  Nav,
-  SSR,
-  Icons,
-  Icon,
-  Overlay,
-  Loading,
-}
+export * from './shared'
+export * from './util'
+export { Nav, OverviewBoxGlobal }
