+++ apps/covid/src/components/shared/Icon.svelte
