+++ apps/covid/src/components/shared/Icons.svelte
