+++ apps/covid/src/components/shared/Loading.svelte
