+++ apps/covid/src/components/shared/Logo.svelte
