+++ apps/covid/src/components/shared/index.js
@@ -1,0 +1,8 @@
+import Icon from './Icon.svelte'
+import Icons from './Icons.svelte'
+import Loading from './Loading.svelte'
+import Logo from './Logo.svelte'
+import Modal from './Modal.svelte'
+import Overlay from './Overlay.svelte'
+
+export {Icon, Icons, Loading, Logo, Modal, Overlay}
