+++ apps/covid/src/components/util/index.js
@@ -1,4 +1,5 @@
 import GetContextFab from './GetContextFab.svelte'
 import RevGeocode from './RevGeocode.svelte'
+import SSR from './styleSSR.svelte'
 
-export { GetContextFab, RevGeocode }
+export { GetContextFab, RevGeocode, SSR }
