+++ apps/covid/src/components/util/styleSSR.svelte
